requirements.txt
structlog
python-dotenv
py-cord